refactor: post header 디자인 시스템 적용
